新编了几个类，加了draw函数

+++ KuroiDrawing/Point.cpp
@@ -1,7 +1,7 @@
 #include "Point.h"
-Point::Point(const double& x_, const double& y_) :x{ x_ }, y{ y_ } {}
+Point::Point(const int& x_, const int& y_) :x{ x_ }, y{ y_ } {}
 Point::Point() : Point{ 1.0,1.0 } {}
-const double& Point::getX() { return this->x; }
-const double& Point::getY() { return this->y; }
-void Point::setX(double& x_) { x = x_; }
-void Point::setY(double& y_) { y = y_; }
+const int& Point::getX() { return this->x; }
+const int& Point::getY() { return this->y; }
+void Point::setX(int& x_) { x = x_; }
+void Point::setY(int& y_) { y = y_; }
